Move files to TeamCode
