--- team_members.md
@@ -1,2 +1,0 @@
-# List of team members
-Ms. Juran
